ai-interviewer/backend/requirements.txt
# FastAPI + Web
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
websockets>=12.0
httpx>=0.26.0

# Pydantic
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
sqlalchemy>=2.0.25
asyncpg>=0.29.0
alembic>=1.13.0
pgvector>=0.2.4

# LLM
litellm>=1.17.0

# Task Queue
celery>=5.3.0
redis>=5.0.0

# Auth
python-jose>=3.3.0
passlib>=1.7.4
google-auth>=2.27.0

# Observability
sentry-sdk>=1.39.0
langfuse>=2.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0

# ASR/TTS
openai>=1.0.0
# elevenlabs>=1.0.0
